Update to inference algorithm
Changed the new inference algorithm to not use the (somewhat problematic)
floor and ceiling operators. The old algorithm is yet to be changed.
Re-wrote the unification rules accordingly. Moved the old rules
(along with the floor and ceil operators into the old/ directory).

Fixed a bug in Unification, inference of polymorphic constraint step:
U({ti=tg})+ had to be changed to U({ti=tg}uC)+