Re-introducing floor and ceil types, fixed heuristic inference.
I realized that we cannot eliminate |_ _| and |- -| types.
The problem is that the soundness theorem must work on *any* consistent
substitution that satisfies the inferred constraints.
Consistency with respect to the base type system cannot be guaranteed
without |_ _| and |- -| types for all substitutions. Hence the
reason to introduce them.

However, this time, the meta-types are not first-class. They are a part of
the function type (syntactically). Functions are written |_t1_| -> |-t2-|
always,  *after* introducing appropriate extra type variables for
copy compatibility. Therefore, unification this time is correct.
The key is that meta-types are *not* used for copy compatibility,
but as an enforcement on the final substitutions.

Fixed (greately simplified) the heiristic inference algoritm.