Stated the direct-soundness theorem. Cleaned up Old proofs.
The new statement and proof of the heuristic inference works in two steps:
i) Prove that the intermediate system -- a declarative system
   with maybe types -- is sound with respect to the core type system.
ii) Prove that inferred types are sound with respect to the core type system.

This formulation is considerable simpler than the older direct proof.
A final pro-forma theorem is stated that establishes soundness of inferred
types wrt the core type system.

Removed old proofs that are not used now.