Introduced Constraint based declarative type system.
I realized that the logical type system that we had for BitC is too weak
to prove the soundness of the equational inference system.
The main problem is that the present system had no constraints in the
type scheme. So, we could not type programs such as:

let f = \x.if x^ () () in
   f (dup true):(ref bool)
   f (dup true):(ref (mutable bool))

Therefore, this commit introduces a new logical type system with
constraints. The soundndess proof needs to be redone with respect
to this system.

I moved the subtype rules into a different figure (and file)
in order to avoid repetetion. The new rules can be found in cdecl.xmli