Started the implementation of Equational inference.
I thought of implementing the equational inference system
(at least in a minimal form) before proeeding with the proof
for two reasons:

i) It was getting tedious to work with some examples due to the large
number of constraints that must be generated. So, I thought of
generating them using the compiler

ii) I wanted to get a sense of how the types come out. I wanted to make sure
there are no obvious problems in the unification rules. It is more tedious to
fix the proofs than the implementation.

The plan is to implement the new inference system and unification for
the subset B language first, and then extend it to all expression of full BitC.

This commit has the infrastructure necessary for the equational inference
system (in TypeEqInfer.cxx). Moved some common routines to TypeCommon.cxx.

Added a command like option --eqinfer to choose the inference algorithm
(TypeChoice.cxx makes the choice of inference).

The changes to large number of files is due to adding the following
kinds of types in kind.def:

subtype: subtype constraint
pcst: the meta-polymorphic constraint
kvar: polymorphic-kind variable
kfix: mono/poly mappings to kvars.