Separated the heuristic inference and Complete inference into different files.
This is in preparation for the change to the logical type system
in order to support complete-inference. I was also planning to add
pairs to the type system to see the effect of non-top-level only
copy compatibility. Since the systems are moving much apart,
I have separated them into two files.