Simpler specification of acyclicity in Unification rules
In the hybrid inference scheme, since we only have equality constraints,
the specification of acyclicity is simplified.
The only check is done in the variable unifies with type rule:

   U({'a = t} u C) | 'a notin t = a |-> t

All the U'() rules are replaced with U() rules everywhere, and the extra
cyclic() check error rule is removed.

This was triggered by an inconstency in the previous specification.
There is actually one kind of "cycle" that can exist in the constraint set.
This is,
  'b|'a = 'a.

This is a legitimate constraint that MUST NOT unify as 'a |-> 'b|'a
but must unify as 'b |-> 'a.

The variable rule and the maybe rules together handle this correctly now.