Fixed handling if array length inference.
Array length is actually a type variable, which ultimnately
resolves to  singleton types. Strictly speaking, we must use type
variables here. Since we do not have singleton types or dependent
types over arrays, we have arrlen in the type record
itself. However, we still need a level of indirection because the
lengths must also be unified during unification [think of
unification calls such as
     U(t1->minimizeMutability() == t2->mininmizeMutability()) ]
Implicitely, an indirection entry with `0' in it is a length
variable. Copy constructor and TypeSpecialize DO NOT copy this
variable deeply.  This "variable" is not subject to
generalization.