Implemented TypeScheme notmalization.
In the presence of PCSTs, the substitution within type schemes is
not capture avoiding. Therefore, there might be some
substitutions into the free type-variables. If so, they can be
removed from the list of type variables.
This also removes redundant constraints, or PCSTs that are
satisfied and no longer need to be present.