Fixed a bug in Type Generalization.
Fixed a worst bug in pattern generalization -- for let/letrec
as an optimization, only those FTVs that are actually in
in the type of an identifier are selected from the letGather's
type scheme and added to an individual type scheme.
Howeer, this would not take into account the free variables in TCC
that can now be legally present and not be a part of the type.

Fix was easy, but the problem took one day to find, unfortunately.