Fixed the handling of Instance environment.
The instance environment of UOCs was being handled differently from
the other environments. At an import/provide, new binding
was being created and all the instances in the `from' environment
were being *appended* to the `to' environment after checking for collosions.
This mechanism is unnecessary and insufficient. The correct thing to do
is that all aliases should point to the same vector of instances, so that
they are consistently checked by the InferInstance() routine. This is
consistent with the handling of other environments.

This comit makes this change, and some necessary corresponding changes
to the polyinstantiator. All test cases are working okay.

In the UnifiedUOC of the polyinstantiator, and RandT() call was happening
at the initialization step (by initUoc() call made by the initUnifiedUoc()
function). This would introduce all prelude symbols and instances without
full qualifications unnecessarily to the unifiedUOC().
This RandT is now removed.

Fixed the symbol resolver to handle at_fqCtr case correctly, symmetric
to the way it is handled by the type checker.