Fixed the pair selection typing.
Contrary to the previous commit, I decided to keep the full type algebra
in the logical type system. The pair typing is fixed by requiring
that only that type (for the pair) be a `p' type rather than a `t' type.
Even though `t' types are not necessary in the logical system, they
are not harmful as far as I can see, and will align better with inferred types.