Fixed the maybe-consistency rule, wrote some more lemmas for the type system.
The consistency predicate for maybe types was wrong since it used =~ and =~=
copy-compatibility relations, which are not defined over maybe types.
Therefore, changed this to use \/ and /\ functions.

Wrote some more substitution lemmas and properties of the <|: relation.

fnxn.bitc contains a test case which pproves that there is a problem
with the unification rules (and implementation) as it is now
  -- topic for next commit