Redefining semantics of maybe-types, fixed typing, inference of assignment.
There was a completeness bug in the type inference of assignment (:=).
The problem is the constraints {t1 = M'a, t1 = 'b|'c, t2 = 'd|'c}
in the := rule. For example:

suppose p:ref(bool x bool)
\x. x^ := p

will be typed as ref(mutable(bool x bool)) because 'a unifies with bool x bool
Now, a function
\x. (x^ := p ; x^.1:=#t) will fail to type check.

This problem was identified in the fnxn.bitc example of the previous commit.

In order to do this, we need more expressive maybe-types.
There are four semantically distinct and important forms:

w: unconstrained, top-level, deel mutability fixed
'a!w: top-level unknown, deep mutability = w
'a|w: top-level, deep mutability variable
M'a|w: top-level fixed to mutability, deep version variable.

It is the lack of the last case that caused wrong typing of assignemnt.

Therefore, added the case to semantics. Fixed, consistency check
of maybe types.

The definiiton of various predicates on types are now simplified.

Changed the definition of subtyping predicate <|: to check
only well-formedness of structure, not consistency of maybe types.

For example, 'a!bool <: int!bool will hold. The consistency is imposed
by an overall consistency constraint |||- on the typing derivation.

Fixed the grammar of types so that mutable constructor can only
appear on unconstrained types.