Fixed subtyping of maybe types, associated lemmas.
The subtyping of maybe types had a problem in the sense that a
maybe type could be a subtype of either a mutable or a immutable type.

The <|: relation was rtying to epresent two lattices at a time:
the subtype mutable vs immutable and
transformability maybe vs concrete type.

Unfortunately these objectives clash and was causing wrong typing derivation
that I had not noticed.

Therefore, fixed this by saying that maybe types can be involved in subtypes
iff 'a|w <|: \/(t) and vice versa, and w'|w <|: t if w' <|:t.

However, still having problems getting the substitution lemma through
across a subtype derivation, which is necessary to prove subject reduction.