Fix stale copyright notices.
We decided two years ago that BitC would be released under a BSD
copyright. It seems that we didn't do the copyright notice update
thoroughly. This commit corrects that oversight.

The bitc.el file (and only that file) must remain under GPL because
it is based on an existing emacs mode that was licensed under GPL.