Finished the type-substitution helper lemma
This lemma proves that for any S' and S such that
dom(S') and ftvs(S) are disjoint, S' |= t, and S |= S'<t>, then
S<S'> |= S<t>.
This lemma is necessary to prove correctness of substitution over
the type instantiation rule. S' will be the substitution for instantiation
from the type scheme, and S is the new substitution over the entire derivation.