Wrote some cases of preservation theorem, type substitution lemma.
I realized that for the preservation theorem to go through,
another lemma is needed. The value-substitution lemma (used by
the main theorem) and the preservation theorem need consistency of
constraints in the declarative derivation. Unlike in the case of
constraint sets, there is no single place to state the consistency
of all types in the derivation. Note that after new substitution rules,
there cannot be a type t|t' such that t ! ==~ t' since we onlt have
'a|t or M'a|t. However, we might have types such as 'a|bool and 'a|int.
Now, in order to express consistency, we can
 --> Thread a set that collects all maybe types in the derivation
     and require that there be a solution to that set.
 --> Express a consistency constraint over the entire derivation
     (not elegant wrt theorem statement)
 --> Show that if there exists a derivation for a consistent type
     under consistent assumptions, then there exists at least ont
     canonical derivation where all types used in the derivation
     are consistent.

I am attmpting the last appproach. This is a commit before beginning
a new lemma that proves the last property.