Definition of consistent type derivation.
After simplification of the syntax of maybe types, we can only have
types of the form 'a!p and w|p. This prevents the occurence of
inconsistent types such as int|bool. However, inconsistencies
of the form 'a|int x 'a|bool can still exist.

It seemed that these might hamper subject reduction. In particular,
the type substitution theorem. Therefore, introduced a notion of
consistent typing derivation
 _                          _
|  ;S |-* e:t, where |||- {| , S , t} at EVERY STEP of derivation.

The main motivation of this deinition is to simplify the type-substitution
lemma, which I thought was necessary in its most generic form --
especially from the point of inference proofs.
But, I think this will be unnecessary since the unification algorithm
is constraint based, and the substitution that solves the final constraints
will solve constraints at all stages. So, a more restrictedlemma
is sufficient at this stage.

More importantly, the abocve form of inconsistencies do not seem to cause
a stuck state because the only thing that the declarative system does
with maybe types is to ignore their maybe-ness.

Therefore, this definition will not be used in further commits.
But, I wanted it to be present in the archive.