Fixed several lemmas in type system proofs, reorganization of proof structure.
1) There existed a lemma in the previous commit that said:
If  S |= Si<t>
    Si |= t
    dom(Si) \inter ftv(S) = 0

  then S<Si> |= S<t>.

This lemma was used to prove the T-ID case in
type substitution lemma, which will be used by value-substitution
and finally by T-Let case of subject reduction.

However, one of the cases of the proof in the above lemma did not
go through. Upon reconsideration, it turns out not to be true.

Here is a counter example:

S = 'c |-> unit
Si = 'a |-> int x bool , 'b |-> ('c!int) x bool
t = 'a | 'b

Fortunately, there is a way to prove the T-ID case without
using this lemma, but using a weaker property.
Wrote these changes to the declarative proofs section.
The T-ID case now successfully goes through.

2) Introduced a formalism for equality under mutability minimization
   (top-level and full copy-level). Introduced certain lemmas to show
   properties of these relationships.

3) Many definitions and lemmas now use the relationship defined in (2).

4) Re-wrote the consistency of substitution under type derivation lemma
   (Now Lemma 15).