Type system with \/-Elimination rule.
Introduced a non-deterministic forall-elimination like rule
into the type system. This is intended to solve the problems with
Type-substituted derivation, by construction.

There is no need for a forall-introduction rule. Simplified T-ID
and T-Let-P rules.

Wrote expressions considered as Syntactic values as a grammar rule,
rather than as a predicate.