Stated consistency of maybe types inductively, strenghtened T-ID rule
There was a problem in the definition of consiistency of
maybe typoes: |||-{t}. Since this is a global property stated without a
reference, it could not be inductively defined. This made the
strengthening lemma over maybe typesimpossible to prove. As a result
type substitution -> value substitution -> preservation were all stuck.

Now stated the consistncy lemma through a reference constraint set.
The final statement of subject reduction is not complicated with
this reference set because of existential quantification over it.

Moved some lemmas that state properties about types and substitutions
into static semantics section.

Strenghtened the pre-requisits of T-ID rule to S |||- t from S |= t
Subject reduction will get stuck without this because of
constraints in {| t |} that may not be presernt in S {| t |}.

Changes to other files containing rules is only that I made it
full size rather than smalll (and moved a few rules to fit the page).