Relaxed type rule for Function values, changed MPC satisfaction.
#1
The type rule for functions was too restrictive due to the
presence of exact equality requirement.

For example, we could not type the following expression:

p:mutable(bool) |-  if true then \x.x else \x.p

As a result, the proof of soundness of inference got stuck.
The current fix should solve the problem.
Propagated the changes to all relevant lemmas and theorems (and proofs)
in the type system.

#2
Made the MPC satisfaction rule stricted by requiring that the
substitution must also make the general type fully immutable
in the polymorphic case and equal in the monomorphic case.