Strengthened the type system wrt meta-polymorphic constraints
The proof of soundness of the inference system was stuck
because the type system was to weak to support it.
The correspondence badly breaks down in the case of let-expressions
since the set of free type variables are not the same anymore.
Therefore, strengthened the base type system to cope with MPC * constraints
Proved subject reduction and progress on this type system.