Rework some inference lemmas, proofs.
Stated the soundness and completeness lemmas of inference directly
now that the type system is more powerful (instead of the intermediate system).

Setup certain helper lemmas. Separated the notions of constraint solving
and constraint satisfaction.

Fixed some small bugs in the unification theory -- the LHS patterns
are written as disjoint unions.

Wrote some more cases of proofs in the inference system.