Revise handling of unit types at parameters and return values.
Functions which return unit are not emitted as returning void. When a
function returning unit is applied, we now emit a comma expression to
restore the missing unit value.

Function parameters of unit type are no longer included in
declarations or definitions, nor are they passed as actual
parameters. Since this omission is occurring after the SSA pass has
been run, any side effects of the associated expressions are still
run. This suppression has no bad effect on (by-ref (mutable ())),
because there is only one possible legal value and the parameter
already contains that.

Because those parameters may still be referenced within the procedure
body, the suppressed parameters are now declared as locals at the top
of the procedure body.