Fix for a null type field access in code generator
In gen-c.cxx, line 1958, isUnitType() was being called on
a AST that had no type field.
All ASTs that are processed by the type checker at least once have a
type field.  The code generator was producing a temporary local
identifier __bitc_temp_mvec in line 2061 in order to get the
at_apply code to do the work for at_makrvector. This AST had no type field,
and it was unnecessary till now.

Added the type record for this identifier. The wchar.c test case compiles
correctly now.