Parser is no longer a pass.
In the previous implementation, the parser ran as the first pass of
Compile(). It is now run prior to calling Compile(). Calls to admit
the resulting UocInfo objects into the ifList and srcList
(respectively) and to call Compile() on them are now made from inside
the parser. Neither the lexer nor the Compile() logic now depend on
the origin being a file. They merely require an origin name string for
diagnostic purposes.

The new approach has two advantages:

  We can compile from things other then files.
  With a minor further tweak, we can now permit multiple (module)
    forms to appear at top level. This allows us to combine multiple
    source files into a single output file without altering name
    visibility rules.

I'm not yet taking advantage of either capability, but that is the
next step.

This won't compile without an update to libsherpa, and I can't log in
to the office remotely at the moment. I'll push the libsherpa update
tomorrow AM and then push this.