Fabricate UocInfo objects from inside the parser, drop parser pass.
The fe_parse pass is now vestigial. It is preserved only for purposes
of diagnostics and tracing.

The UocInfo objects are now generated from inside the parser. This
will allow us to handle multiple (module...) forms originating in a single
source unit of compilation. It should also somewhat simplify making an
incremental interactive compiler.

The top ast node of a UocInfo is now an at_module or at_interface AST.
I have dropped the at_start and the at_bitc_version node types. Checking
of (bitc-version x) is now done in the parser, and the bitc-version node
is not present in the AST that gets attached to the UocInfo.