Made changes to error reporting in some cases
Some cases in the type inference engine were still using the
AST field in the type record. Changes this to explicitely pass/use
an AST location for error reporting

This is in preparation for removing AST fiedld in type records.

Added a stdout file for twomodule to mathc expected output.