Removed the (now unused) AST backpointer from Type structure.
The AST backpointer was a NoGcPtr to AST. This was
  i) Creating a Null reference problem due to some temporary
     type record beeing created with a collectable AST
  ii) Not effective for error reporting due to linkage of type records.

Therefore, removed this field from the type record. The error reporting
is now handled by explicit bookkeeping wherever necessary.

The Unifier, solver, etc were recieving an AST pointer for retrieving the
AST location where the error must be reported. Changed these routines
to instead accept a reference to the lexical location where the
error (if any) must be reported.

All test cases in the unit directory are now working without errors.