Turn Options::entryPts into a set.
It turns out that the old entryPts list was not precisely a set,
because it had an ordering dependency: bitc.prelude.__index_lt and
bitc.prelude.IndexBoundsError needed to be instantiated before
anything else, or there would be unresolved symbols.

To solve that, I moved the instantiation of those symbols to
UocInfo::instantiateBatch (Instantiate.cxx), where they are now done
by hand. If we decide to do an interactive compiler, these will need
to be instantiated from bitcc.cxx again.

I would have simply called InstantiateBatch multiple times from
bitcc.cxx, but I wasn't convinced that it was serially
re-entrant. Need to check with Swaroop on this.