Switch to a non-invasive counting pointer implementation.
This version remembers tracked pointers so that it can re-associate
the correct counter object. It is structurally similar to the boost
implementation, but I'm not quite ready to cut over to that. The goal
here is to let me remove the last vestiges of CVector.

This implementation is significantly slower than the old one. Cannot
be helped given the need to associate a counter object.