GCPtr now valid on forward-declared classes, fixes to NoGCPtr, operator=
This variant uses the same basic trick as the boost libraries to permit
declartion of reference-counted pointers to classes that are not yet
defined.

The NoGCPtr class no longer admits dereference, comparison, or boolean test,
because its value is not stable. To do that, first copy the NoGCPtr to a
GCPtr and then test that. Minor update in Symtab.cxx reflecting that.

Inserted a fix (and associated comment) into GCPtr assignment operator to
deal with the case where assignment causes destruction of the target GCPtr
(which has been passed to the assignment operator by reference).