Fixed a problem with overlap-checking for typeclass instances.
Fixed the overlap-checking in InferInstance to use
Instance::overlaps() -- check for type unifiability rather than
Instance::equals() -- check type unifiability, and
                      solvability of all preconditions

This solves the problem with compilation error in disjoint.bitc:

In disjoint.bitc we had two instances for the class ABC:

 (definstance (forall ((IntLit 'a)) (ABC 'a))  ... )
 (definstance (forall ((FloatLit 'a)) (ABC 'a)) ...)

If we look at the type of the instance being defined, both have type
(ABC 'a), and therefore are overlapping.

However, at the point of  inference, these were deemed acceptable since
we are using Instance::equal() for comparing two instances
[Instance::equal() returns true only if the types unify and all
pre-conditions match]. This is actually an insufficient overlap-check
because if FloatLit later has an instance (FloatLit int32) (or if IntLit
has some other instance that is also an instance of FloatLit), we have
lost the overlapping restriction. In this case, we know that FloatLit
and IntLit should not overlap, but this is not true in general.
In the Solver, what happened is that for a constraint (ABC int32)
1) First came across the (((FloatLit 'a)) (ABC 'a)) instance
2) Deemed 'a unifiable with int32, assuming non-overlap
3) Added pre-condition (FloatLit int32) which is unsatisfiable.

The solver unifies with the first unifiable instance. It is not a
backtracking solver which tries other instances if the overall
solving fails for an instance (here (FloatLit 'a)).

Previously when Instance-list was a vector, we would see the
(((IntLit 'a)) (ABC 'a)) constraint first and solve all constraints.

The overlapping instances in disjoint.bitc have been removed.
The original code for disjoint.bitc has now moved to test/error, along
with an explanation. Removed file inst.bitc, which was essentially
the same as disjoint.bitc.