Specified type equivalence for constrained types.
In the case of constrained types 'a|t1 and 'a|t2, defined a notion
of equivalence when \/(t1) = \/(t2).

The types 'a|t1 and 'a|t2 are structurally different but are
semantically equivalent.

In the unifier, we actually produce solutions to types that are
equivalent (that is, = constraints are actually treated as
equivalence constraints) See U-Ct1 and U-Ct3 rules.

Further the subtyping relationship must deal with function types
and refeernce types wrt equivalence, not structural equality.
otherwise, we cannot show soundness of unification.
This is because, the unifier will unify types such as
'a|t1 -> () = 'a|Mt1 -> ()
which must also be recognized in the logical type system.

Defined semantic equivalence of types and noted that it is
identified as equivalent to the structural equality relation.