Better handling of identifier types in ASTs and Symbol resolver.
The previous identType scheme was complicated die to the presence of
certain identifiers that perform dual roles
(ex: structure name is a type as well as a constructor).
In order to cope with this, we had an arrangement that involved both
identType as well as flags, which created some inelegance.

This commit fixes the above problem by using a identType grouping scheme
similar to ASTmaker.

We noe have the following identifier types:

  identType     Meaning

  id_tvar       Type variables
  id_union      Union/repr name
  id_struct     Structure type name
  id_typeclass  Type class name
  id_method     type class method name
  id_field      Structure or union field name
  id_interface  interface name (local name in the current source unit)
  id_value      Any identifier defined at define, let, lambda,
                do, switch, case.
  id_ucon       Union constructor having >=1 argument
  id_ucon0      Union constructor taking no arguments

We also have the following groupings:

Types:
  idc_type      id_tvar || id_union || id_struct

Values:
  idc_value     id_value || id_enumcon || id_method

Union Constructors:
  idc_uctor     id_ucon || id_ucon0

All Constructors:
  idc_ctor      id_struct || id_ucon || id_ucon0

Identifiers legal at apply-position:
  idc_apply      idc_value || idc_ctor