Changed code generator to handle proclaimations locally
The code generator was using an AST flag to track proclaimations that
are already emmitted, and thus avoid emitting them again. This is only a
concern in the case of types as we emit a typedef, not just a forward
declaration.

Replaced the code to instead use a set<string> that is private to the
code generator, and removed the assiciated AST flag.