Updated Unification rules for const, composite-mutable handling
Added new const-rule to unification. There is a need for normalizing types
so that const is lazily propagated inwards, and the unification algorithm
only sees const over type variables. Normalization is also necessary
to ensure that the mutability of a composite type is defined by
(and during inference influences) the mutability of inner unboxed fields.

The actual normalization rules still need to be defined. The present
unification algorithm assumes (and uses normalization).