Fixed Unification wrt Mutability propagation.
Added a M-function which propagates mutability inwards, and
used it in the U-Mut rule.

Mutability propagation is handled at the mutability
unification rule, which is the only place where new mutability
can be generated.

It is sufficient to handle this propagation at the U-Mut rule
since all types that share sub-substructure are changed
automatically (due to substitutions created by M-function), and
mutability only needs to be propagated inwards.

Completeness: It is the responsibility of the inference
algorithm (and the programmer in the case of type annotations)
to produce inner types that are either mutable or accept
mutability. Otherwise, this will lead to unification failure due
to the M-function returning _|_

Decidability: The M-function introduces new type variables,
which is uncharestetistic of unification. However, this is okay
because: (1) M-function does not call unification recursively.
(2) The total number of type variables after the call to
M-function is non-increasing - it only substitutes types to
variables, or new variables to old variables.