Implemented the Mutability propagation function (not used by unifier).
Implemented the mutability propagation function that ensures that
the mutability of an unboxed product is no greater than the mutability
of the composing fields.

This is the extension of the M function (Definition 2 of
 doc/docs/bitc/new-formal.txml) to full BitC type system.

The unifier does not use this function as of now, so the behavior of the
compiler itself is unchanged.