Implemented M'a|t form of maybe types, fixed inference of assignment.
The theory of type inference uses two forms of maybe types:

   mbTop  ::= 'a!p
   mbFull ::=  s|p,

  where s ::= 'a | M'a
    and p c t | p != mbTop and p != mbFull.

However, the implementation only supported the maybe types
'a!p and 'a|p. This implementation is for an older version
of the inference algorithm.

Therefore, implemented the semantics of the M'a|p form of maybe types
and updated the inference and unification algorithm to match the theory.

Made necessary updates to the polyinstantiator to cope with M'a|p types
since unresolved types of this form must be coerced to mutable variants.