Fixed some problems in the complete-tech report due to inclusions.
I had tried to share certain parts of the paper and tech report.
This had some problems such as: the papers says ` detaled proofs
can be obtained in the technical report'' and cites it.
But, the technical report cannot cite itself.  So, gave up some
sharing to fix the problems in the technical report.

Also re-numbered the definitions, emmas and theorems in the text part
(not appendix) of the tech-report in roman numbers so that they do
not collide with the appendix.