Implementation of labeled expressions and labeled escape.
This implements (LABEL ident e) and (RETURN-FROM ident e), which is
comparable to the LISP constructs BLOCK and RETURN-WITH_VALUE. The
choice of different keywords is a bug that will shortly be fixed.

There is, as yet, no implementation of the convenience syntax (RETURN e),
which is a sugaring over labeled expressions.

This is not yet documented in the spec, and there are no test cases yet.

This commit also removes the (wrong) assessment that exceptions
inherently induce the value restriction.