Fixed a symbol-resolver bug wrt incompleteness handling
The incompleteness check in the symbol resolver was unfoortunately
not at all excersized due to a bug in paranthesization. Fixed this bug
(paranthesization) thus enabling incompleteness checking, and some other errors
discovered in the process.

Lambda parameters were not marked complete, fixed this.

In the Environment structure, the name is redundantly encoded in to places:
as the key to the Bindings map
as the nm field in the bindings structure.

This lead to some inconsistency due to updateKey() call in the instantiator.
Fixed all uses of the environment to obtain the string(key) from the first component
of the bindings map only. The nm field in the binding-structure (I think) should
really be retired at this point.