Fixed theory, implementation of const, unification under maybe.
The rule for unification under maybe-types or const was one that is
performed under the minimizeMutability (\bigtriangledown) operator
that removes mutability up-to the ref/function boundary.
However, this operator preserves const-ness around variable types,
which is not the desirable rule under this unification because it leads to
incompleteness. For example,

|_ |_'c|'b_| x bool _| == |_ 'd|'b x bool _|

will fail under the rule. The issue to note here is that under the
const operator (or on the RHS of a mbFull type), the mutability
of the type itself doe not matter as the outer const will
anyway remove it (or we can get any equivalent type in the case of mbFull).

So, we need not preserve the const-ness in these cases. We need an operator
that minimizesMutability by eagerly interpreting the const meta-constructor.
This function is implemented as
minMutConstless() [\mathaccent\circ\bigtriangledown]
and is used by the type rules and unification routines.

The copy-coercion rules need not use this operator since they operate
one step at a time and will eventually interpret the const meta costructors
stepwise.