Halfway to structure methods.
Augmented at_usesel to handle structure methods S.m.

Resolver now recognizes methods. Type checker also, and adds them to a
new methods[] vector in the type record.

Method select is (for the moment) only valid in at_apply in
applicative position. This is handled in the type checker, where the
(s.m x y) is re-written into a function call (S.m s x y).

Type checker is not yet typing method implementations or inserting the
appropriate proclaim when it encounters a method. That is next.