Fixed type instantiation, definition/declaration matching.
The function mtchDeDecl() which checks strict compatibility of definitions
with declarations, was receiving a parameter fnCopyCompatibility.
When this parameter was true, the function types were compared with
minimizeMutability() on args and return types.
This step was both unnecessary (since functions have a canonical representation
in the type system) and incorrect (since only outermost function type
was checked under this condition).

Removing this logic resulted in failure of the SHA1-test cases. Ultimately,
I found that the real problem was in the Instantiate() function in
TypeInfer. This function would selectively call type_instance_copy()
and type_instance() based on whether ftvs were non-empty/empty.
In cases where the function type is fully concrete, this will return the
type within the type scheme. This clashes with the function type normalization
because fixupFnTypes() will actually change the type within the
typeScheme itself. Therefore, the type checking was failing when
run for the second time (actually, second instantiation).

Fixed this, so that Instantiate() always instantiates by copy.
The function (formerly called) type_instance() is no longer necessary,
and thus removed it. Former type_instance_copy() is not type_instance()
Similarly for ts_instance().