Repair separator handling for imported bindings.
At some point I decided that there should be a visible external distinction
between an FQN (ifname:id) and a local binding of an imported name
(ifBinding.id). To avoid changing billions of references in the future, I
added an FQName::sep constant for the FQN separator and then went through
and replaced a lot of uses of '.' with FQName::sep. In the process, I
mistakenly changed the three places where local bindings were being fabricated.
This broke the pretty-printer output, which caused breakage when anything
was linked against the standard library. Because our test environment
uses --nostdlib, we didn't catch this.

I have now added FQName::LocalBindingSep (which is '.') and revised those
three places to use the correct separator. This resolves Martin's bug,
but I still need to look at how to add a test case that will ensure that
we do not regress. Before I do that, I want to deal with another cleanup
issue.