Updates for powerpc provided by Paul Snively.
Apparently the listvec test is not working, and it isn't clear
why. The moral of this patch, overall, is that we need to bite the
bullet and deal with extern C calling convention in a first-class way,
and deal with closure-based calling conventions using a bitc-specific
calling convention. On many platforms it will turn out that this
matches the native calling convention (or is at least compatible), but
on PPC that appears not to be the case.