Retire AST::isLocation()
The isLocation() method was originally introduced to let us check
whether the LHS of an assignment was either (a) a reference type, or
(b) something other than a temporary. That is: it provided the classic
l-value check.

Responsibility for this had long since been taken over by LocChk.cxx
(which, I note, doesn't handle array-ref yet). I'm retiring the old
one so that we aren't maintaining vestigial code, but also because it
wasn't clear what it should do for generalized nth/length.