First, halting steps toward new syntax.
Updated the version number.

Added initial implementation of "bare" NTH in TypeInfer.cxx, AST.*

Revised comment syntax to accept // and /* style comments unconditionally.

Updated the lexer to modally accept new syntax tokens, old syntax
tokens, or both (prior to the version statement) so that we can build
a transitional parser.

Parser now accepts the version form in the new syntax, and if present,
does NOT treat ";" as a comment-start character after the latest
version is seen. In effect, the new-syntax version statement signals
the beginning of input in the new syntax, and a bunch of the lexing
rules change accordingly.

Fixed an output bug discovered in Type-debug-pp.cxx and Type-pp.cxx.