Vector, array, and array-ref now have a .length field by courtesy.
Simulate the presence of a .length field in all three types, to
properly infer has-field over the use cases, and to add special
handling of these types in the constraint solver. Note this has the
somewhat curious property that all arrays are part of the

   (has-field 'a length Word)

class even though we can't (in general) abstract over array size
anywhere else.